requirements/local.txt
asgiref==3.7.2
Django==4.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-model-utils==4.3.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
Unipath==1.1